registry/src/requirements.txt
fastapi
uvicorn
typer
requests
boto3
python-multipart
python-dotenv
